Move `{LintDocs,TierCheck}` into `doc_tool_tests`
