+++ src/bootstrap/src/core/build_steps/test/mod.rs
@@ -65,92 +65,9 @@ pub(crate) use compiletest_suites::{
 };
 pub(crate) use devtool_tests::{Cargo, Clippy, RustAnalyzer, Rustfmt};
 pub(crate) use dist_tool_tests::{CollectLicenseMetadata, Distcheck, RustInstaller};
-pub(crate) use doc_tool_tests::{HtmlCheck, Linkcheck};
+pub(crate) use doc_tool_tests::{HtmlCheck, Linkcheck, LintDocs, TierCheck};
 pub(crate) use miri_tests::{CargoMiri, Miri};
 pub(crate) use rustdoc_tests::{CrateRustdoc, CrateRustdocJsonTypes, RustdocTheme};
 pub(crate) use std_tests::TestFloatParse;
 pub(crate) use test_helpers::Crate;
 pub(crate) use tidy::Tidy;
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub struct TierCheck {
-    pub compiler: Compiler,
-}
-
-impl Step for TierCheck {
-    type Output = ();
-    const DEFAULT: bool = true;
-    const ONLY_HOSTS: bool = true;
-
-    fn should_run(run: ShouldRun<'_>) -> ShouldRun<'_> {
-        run.path("src/tools/tier-check")
-    }
-
-    fn make_run(run: RunConfig<'_>) {
-        let compiler =
-            run.builder.compiler_for(run.builder.top_stage, run.builder.build.build, run.target);
-        run.builder.ensure(TierCheck { compiler });
-    }
-
-    /// Tests the Platform Support page in the rustc book.
-    fn run(self, builder: &Builder<'_>) {
-        builder.ensure(compile::Std::new(self.compiler, self.compiler.host));
-        let mut cargo = tool::prepare_tool_cargo(
-            builder,
-            self.compiler,
-            Mode::ToolStd,
-            self.compiler.host,
-            Kind::Run,
-            "src/tools/tier-check",
-            SourceType::InTree,
-            &[],
-        );
-        cargo.arg(builder.src.join("src/doc/rustc/src/platform-support.md"));
-        cargo.arg(builder.rustc(self.compiler));
-        if builder.is_verbose() {
-            cargo.arg("--verbose");
-        }
-
-        let _guard = builder.msg(
-            Kind::Test,
-            self.compiler.stage,
-            "platform support check",
-            self.compiler.host,
-            self.compiler.host,
-        );
-        BootstrapCommand::from(cargo).delay_failure().run(builder);
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub struct LintDocs {
-    pub compiler: Compiler,
-    pub target: TargetSelection,
-}
-
-impl Step for LintDocs {
-    type Output = ();
-    const DEFAULT: bool = true;
-    const ONLY_HOSTS: bool = true;
-
-    fn should_run(run: ShouldRun<'_>) -> ShouldRun<'_> {
-        run.path("src/tools/lint-docs")
-    }
-
-    fn make_run(run: RunConfig<'_>) {
-        run.builder.ensure(LintDocs {
-            compiler: run.builder.compiler(run.builder.top_stage, run.builder.config.build),
-            target: run.target,
-        });
-    }
-
-    /// Tests that the lint examples in the rustc book generate the correct lints and have the
-    /// expected format.
-    fn run(self, builder: &Builder<'_>) {
-        builder.ensure(crate::core::build_steps::doc::RustcBook {
-            compiler: self.compiler,
-            target: self.target,
-            validate: true,
-        });
-    }
-}
